// src/pages/enums/ECity.enum.tsx
export enum ECity {
    ARMAZEM = 'Armazem',
    BRACO_DO_NORTE = 'BracodoNorte',
    CAPIVARI_DE_BAIXO = 'CapivarideBaixo',
    GAROPABA = 'Garopaba',
    GRAO_PARA = 'GraoPara',
    GRAVATAL = 'Gravatal',
    IMARUI = 'Imarui',
    IMBITUBA = 'Imbituba',
    JAGUARUNA = 'Jaguaruna',
    LAGUNA = 'Laguna',
    ORLEANS = 'Orleans',
    PEDRAS_GRANDES = 'PedrasGrandes',
    PESCARIA_BRAVA = 'PescariaBrava',
    RIO_FORTUNA = 'RioFortuna',
    SANGAO = 'Sangao',
    SANTA_ROSA_DE_LIMA = 'SantaRosadeLima',
    SAO_LUDGERO = 'SaoLudgero',
    SAO_MARTINHO = 'SaoMartinho',
    TREZE_DE_MAIO = 'TrezedeMaio',
    TUBARAO = 'Tubarao',
}

export enum ECityTranslate {
    Armazem = 'Armazém',
    BracodoNorte = 'Braço do Norte',
    CapivarideBaixo = 'Capivari de Baixo',
    Garopaba = 'Garopaba',
    GraoPara = 'Grão-Pará',
    Gravatal = 'Gravatal',
    Imarui = 'Imaruí',
    Imbituba = 'Imbituba',
    Jaguaruna = 'Jaguaruna',
    Laguna = 'Laguna',
    Orleans = 'Orleans',
    PedrasGrandes = 'Pedras Grandes',
    PescariaBrava = 'Pescaria Brava',
    RioFortuna = 'Rio Fortuna',
    Sangao = 'Sangão',
    SantaRosadeLima = 'Santa Rosa de Lima',
    SaoLudgero = 'São Ludgero',
    SaoMartinho = 'São Martinho',
    TrezedeMaio = 'Treze de Maio',
    Tubarao = 'Tubarão',
}

export default ECity;