// src/pages/enums/EVehicles.enum.tsx
export enum EVehicles {
    TRUCK = 'caminhao',
    BUS = 'onibus',
    CAR = 'carro',
    MICROBUS = 'micro-onibus',
    BIKE = 'moto',
}

export enum EVehiclesTranslate {
    caminhao = 'Caminhão',
    onibus = 'Ônibus',
    carro = 'Carro',
    microonibus = 'Micro-ônibus',
    moto = 'Moto',
}

export default EVehicles;